+++ src/config.ts
@@ -63,7 +63,7 @@ export function defaultConfig(
     fileFormat: 'json',
 
     // Use the marker defaults value? Default is true
-    userMarkerDefaults: true,
+    useMarkerDefaults: true,
 
     // Set a default-language for the marke default value. Default is 'en'
     defaultLanguage: 'en'