+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-## [1.0.0] (UNRELEASED)
+## [1.0.0] (01/03/2024)
 
 ### Changes 
 * Removed "missing value" string as default for translations. 
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file. See [standa
 * Added cli and settings options for "useMarkerDefaults" and "defaultLanguage" parameters.
 * Added "workdir" parameter to set the working directory
 * Updated marker function to allow a translation value for the default language (set on extraction).
+* Added extraction of default values set in code. 
  
 ### Init
 * Forked "jsverse/transloco-keys-manager" on 2024-02-29.